services/market-data-collector/requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
ccxt==4.5.44
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
coincurve==21.0.0
colorama==0.4.6
cryptography==46.0.5
fastapi==0.135.2
frozenlist==1.8.0
h11==0.16.0
idna==3.11
multidict==6.7.1
numpy==2.4.3
pandas==3.0.1
propcache==0.4.1
pycares==5.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
setuptools==82.0.1
six==1.17.0
starlette==1.0.0
ta==0.11.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
yarl==1.23.0
prometheus-fastapi-instrumentator 
grpcio 
protobuf 
prometheus-fastapi-instrumentator 
